move publishing utils to separate file, and disable CC because of POM hackery
